Created the Base Application
Made a Base Application with one activity.
Potential Issues:
I have linked Git to it as a version control so not sure if that will
carry over to another pc - if it does I'll release the git linking
I have created a virtual device that doesn't want to work on my laptop -
I'm not sure if it will carry over, and if it does, whether it will
work.