Made a good start
I added all the pages we need to make.
Added some 'fragments' that we can use to put two on one page like he
wanted us to, then linked one to the topics page so all appears normal.
Added Scrollviewer around the topic fragment to allow scrolling in
landscape
Set it to remove the ActionBar from the page when it is rotated to
landscape - this means more screen real estate.
Removed the ActionBar from the landing page.
Made the landing page background full screen